+++ packages/validator-ajv8/src/validator.ts
@@ -1,35 +1,30 @@
 import Ajv, { ErrorObject, ValidateFunction } from 'ajv';
-import toPath from 'lodash/toPath';
-import isPlainObject from 'lodash/isPlainObject';
-import clone from 'lodash/clone';
+import get from 'lodash/get';
 import {
+  createErrorHandler,
   CustomValidator,
-  ERRORS_KEY,
   ErrorSchema,
-  ErrorSchemaBuilder,
   ErrorTransformer,
-  FieldValidation,
   FormContextType,
-  FormValidation,
-  GenericObjectType,
   getDefaultFormState,
-  mergeValidationData,
-  REF_KEY,
+  getUiOptions,
+  PROPERTIES_KEY,
   RJSFSchema,
   RJSFValidationError,
+  ROOT_SCHEMA_PREFIX,
   StrictRJSFSchema,
+  toErrorList,
+  toErrorSchema,
   UiSchema,
+  unwrapErrorHandler,
   ValidationData,
+  validationDataMerge,
   ValidatorType,
-  PROPERTIES_KEY,
-  getUiOptions,
+  withIdRefPrefix,
 } from '@rjsf/utils';
-import get from 'lodash/get';
 
 import { CustomValidatorOptionsType, Localizer } from './types';
 import createAjvInstance from './createAjvInstance';
-
-const ROOT_SCHEMA_PREFIX = '__rjsf_rootSchema';
 
 /** `ValidatorType` implementation that uses the AJV 8 validation mechanism.
  */
@@ -59,133 +54,20 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
     this.localizer = localizer;
   }
 
-  /** Transforms a ajv validation errors list:
-   * [
-   *   {property: '.level1.level2[2].level3', message: 'err a'},
-   *   {property: '.level1.level2[2].level3', message: 'err b'},
-   *   {property: '.level1.level2[4].level3', message: 'err b'},
-   * ]
-   * Into an error tree:
-   * {
-   *   level1: {
-   *     level2: {
-   *       2: {level3: {errors: ['err a', 'err b']}},
-   *       4: {level3: {errors: ['err b']}},
-   *     }
-   *   }
-   * };
-   *
-   * @param errors - The list of RJSFValidationError objects
-   * @private
-   */
-  private toErrorSchema(errors: RJSFValidationError[]): ErrorSchema<T> {
-    const builder = new ErrorSchemaBuilder<T>();
-    if (errors.length) {
-      errors.forEach((error) => {
-        const { property, message } = error;
-        const path = toPath(property);
-
-        // If the property is at the root (.level1) then toPath creates
-        // an empty array element at the first index. Remove it.
-        if (path.length > 0 && path[0] === '') {
-          path.splice(0, 1);
-        }
-        if (message) {
-          builder.addErrors(message, path);
-        }
-      });
-    }
-    return builder.ErrorSchema;
-  }
-
   /** Converts an `errorSchema` into a list of `RJSFValidationErrors`
    *
    * @param errorSchema - The `ErrorSchema` instance to convert
    * @param [fieldPath=[]] - The current field path, defaults to [] if not specified
    */
   toErrorList(errorSchema?: ErrorSchema<T>, fieldPath: string[] = []) {
-    if (!errorSchema) {
-      return [];
-    }
-    let errorList: RJSFValidationError[] = [];
-    if (ERRORS_KEY in errorSchema) {
-      errorList = errorList.concat(
-        errorSchema[ERRORS_KEY]!.map((message: string) => {
-          const property = `.${fieldPath.join('.')}`;
-          return {
-            property,
-            message,
-            stack: `${property} ${message}`,
-          };
-        })
-      );
-    }
-    return Object.keys(errorSchema).reduce((acc, key) => {
-      if (key !== ERRORS_KEY) {
-        const childSchema = (errorSchema as GenericObjectType)[key];
-        if (isPlainObject(childSchema)) {
-          acc = acc.concat(this.toErrorList(childSchema, [...fieldPath, key]));
-        }
-      }
-      return acc;
-    }, errorList);
-  }
-
-  /** Given a `formData` object, recursively creates a `FormValidation` error handling structure around it
-   *
-   * @param formData - The form data around which the error handler is created
-   * @private
-   */
-  private createErrorHandler(formData: T): FormValidation<T> {
-    const handler: FieldValidation = {
-      // We store the list of errors for this node in a property named __errors
-      // to avoid name collision with a possible sub schema field named
-      // 'errors' (see `utils.toErrorSchema`).
-      __errors: [],
-      addError(message: string) {
-        this.__errors!.push(message);
-      },
-    };
-    if (Array.isArray(formData)) {
-      return formData.reduce((acc, value, key) => {
-        return { ...acc, [key]: this.createErrorHandler(value) };
-      }, handler);
-    }
-    if (isPlainObject(formData)) {
-      const formObject: GenericObjectType = formData as GenericObjectType;
-      return Object.keys(formObject).reduce((acc, key) => {
-        return { ...acc, [key]: this.createErrorHandler(formObject[key]) };
-      }, handler as FormValidation<T>);
-    }
-    return handler as FormValidation<T>;
-  }
-
-  /** Unwraps the `errorHandler` structure into the associated `ErrorSchema`, stripping the `addError` functions from it
-   *
-   * @param errorHandler - The `FormValidation` error handling structure
-   * @private
-   */
-  private unwrapErrorHandler(errorHandler: FormValidation<T>): ErrorSchema<T> {
-    return Object.keys(errorHandler).reduce((acc, key) => {
-      if (key === 'addError') {
-        return acc;
-      } else {
-        const childSchema = (errorHandler as GenericObjectType)[key];
-        if (isPlainObject(childSchema)) {
-          return {
-            ...acc,
-            [key]: this.unwrapErrorHandler(childSchema),
-          };
-        }
-        return { ...acc, [key]: childSchema };
-      }
-    }, {} as ErrorSchema<T>);
+    return toErrorList(errorSchema, fieldPath);
   }
 
   /** Transforming the error output from ajv to format used by @rjsf/utils.
    * At some point, components should be updated to support ajv.
    *
    * @param errors - The list of AJV errors to convert to `RJSFValidationErrors`
+   * @param [uiSchema] - An optional uiSchema that is passed to `transformErrors` and `customValidate`
    * @protected
    */
   protected transformRJSFValidationErrors(
@@ -215,7 +97,7 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
 
         stack = message;
       } else {
-        const uiSchemaTitle = getUiOptions(get(uiSchema, `${property.replace(/^\./, '')}`)).title;
+        const uiSchemaTitle = getUiOptions<T, S, F>(get(uiSchema, `${property.replace(/^\./, '')}`)).title;
 
         if (uiSchemaTitle) {
           stack = `'${uiSchemaTitle}' ${message}`.trim();
@@ -307,7 +189,7 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
       errors = transformErrors(errors, uiSchema);
     }
 
-    let errorSchema = this.toErrorSchema(errors);
+    let errorSchema = toErrorSchema<T>(errors);
 
     if (invalidSchemaError) {
       errorSchema = {
@@ -325,41 +207,9 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
     // Include form data with undefined values, which is required for custom validation.
     const newFormData = getDefaultFormState<T, S, F>(this, schema, formData, schema, true) as T;
 
-    const errorHandler = customValidate(newFormData, this.createErrorHandler(newFormData), uiSchema);
-    const userErrorSchema = this.unwrapErrorHandler(errorHandler);
-    return mergeValidationData<T, S, F>(this, { errors, errorSchema }, userErrorSchema);
-  }
-
-  /** Takes a `node` object and transforms any contained `$ref` node variables with a prefix, recursively calling
-   * `withIdRefPrefix` for any other elements.
-   *
-   * @param node - The object node to which a ROOT_SCHEMA_PREFIX is added when a REF_KEY is part of it
-   * @private
-   */
-  private withIdRefPrefixObject(node: S) {
-    for (const key in node) {
-      const realObj: GenericObjectType = node;
-      const value = realObj[key];
-      if (key === REF_KEY && typeof value === 'string' && value.startsWith('#')) {
-        realObj[key] = ROOT_SCHEMA_PREFIX + value;
-      } else {
-        realObj[key] = this.withIdRefPrefix(value);
-      }
-    }
-    return node;
-  }
-
-  /** Takes a `node` object list and transforms any contained `$ref` node variables with a prefix, recursively calling
-   * `withIdRefPrefix` for any other elements.
-   *
-   * @param node - The list of object nodes to which a ROOT_SCHEMA_PREFIX is added when a REF_KEY is part of it
-   * @private
-   */
-  private withIdRefPrefixArray(node: S[]): S[] {
-    for (let i = 0; i < node.length; i++) {
-      node[i] = this.withIdRefPrefix(node[i]) as S;
-    }
-    return node;
+    const errorHandler = customValidate(newFormData, createErrorHandler<T>(newFormData), uiSchema);
+    const userErrorSchema = unwrapErrorHandler<T>(errorHandler);
+    return validationDataMerge<T>({ errors, errorSchema }, userErrorSchema);
   }
 
   /** Validates data against a schema, returning true if the data is valid, or
@@ -380,7 +230,7 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
       if (this.ajv.getSchema(rootSchemaId) === undefined) {
         this.ajv.addSchema(rootSchema, rootSchemaId);
       }
-      const schemaWithIdRefPrefix = this.withIdRefPrefix(schema) as S;
+      const schemaWithIdRefPrefix = withIdRefPrefix<S>(schema) as S;
       let compiledValidator: ValidateFunction | undefined;
       if (schemaWithIdRefPrefix['$id']) {
         compiledValidator = this.ajv.getSchema(schemaWithIdRefPrefix['$id']);
@@ -398,21 +248,5 @@ export default class AJV8Validator<T = any, S extends StrictRJSFSchema = RJSFSch
       // make sure we remove the rootSchema from the global ajv instance
       this.ajv.removeSchema(rootSchemaId);
     }
-  }
-
-  /** Recursively prefixes all $ref's in a schema with `ROOT_SCHEMA_PREFIX`
-   * This is used in isValid to make references to the rootSchema
-   *
-   * @param schemaNode - The object node to which a ROOT_SCHEMA_PREFIX is added when a REF_KEY is part of it
-   * @protected
-   */
-  protected withIdRefPrefix(schemaNode: S | S[]): S | S[] {
-    if (Array.isArray(schemaNode)) {
-      return this.withIdRefPrefixArray([...schemaNode]);
-    }
-    if (isPlainObject(schemaNode)) {
-      return this.withIdRefPrefixObject(clone<S>(schemaNode));
-    }
-    return schemaNode;
   }
 }